% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casecross.R
\name{casecross}
\alias{casecross}
\title{Time-stratified case-crossover}
\usage{
casecross(data, id, date, covariate = F, period = "month")
}
\arguments{
\item{data}{Case-only observations you wish to create referent observations for}

\item{id}{String of variable of strata identifier. Leave blank if you do not have a known identifier.}

\item{date}{Date of admission. This is also sets the outcome variable to 1}

\item{covariate}{Strings of covariates you'd like to include in the time-stratified dataframe}

\item{period}{Character value based on lubridates floor and ceiling unit. Default is "month".}
}
\description{
Function creates a time-stratified case-crossover dataframe from a case-only
dataframe where the outcome on a particular date can be compared to referent
periods on the same day of the week.
}
\examples{
casecross()

}
\keyword{case-crossover}
\keyword{case-crossover,}
\keyword{time-stratified}
