% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.stratified.R
\name{time.stratified}
\alias{time.stratified}
\title{Time-stratified case-crossover}
\usage{
\method{time}{stratified}(data, id, covariate = F, admit_date, start_date,
  end_date, interval)
}
\arguments{
\item{data}{Case-only data you wish to create referent observations for}

\item{id}{String of variable of strata identifier. Leave blank if you do not have a known identifier.}

\item{covariate}{Strings of covariates you'd like to include in the time-stratified dataframe}

\item{admit_date}{Date of admission. This is also sets the outcome variable to 1}

\item{start_date}{Start date of the period you wish to create referent observations over}

\item{end_date}{End date of the period you with to create referent observations over}

\item{interval}{Numberic value. Day == 1, Week = 7, Bi-weekly == 14, etc...}
}
\description{
Function creates a time-stratified case-crossover dataframe from a case-only
dataframe where the outcome on a particular date can be compared to referent
periods on the same day of the week.
}
\examples{
time.stratified()
}
\keyword{case-crossover}
\keyword{case-crossover,}
\keyword{time-stratified}
